refactor: Registration Steps separated into different components under its file step
