+++ src/app/registration/[eventId]/_components/forms/RegistrationPayment.tsx
